+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Version 6
 
+## [6.5.7] - 2024/06/19
+### Changed:
+- 9x8x8x9 model architecture
+- smart training
+
+## [6.5.6] - 2024/06/19
+### Changed:
+- 9x15x9 model architecture
+- smart training
+
 ## [6.4.5] - 2024/06/19
 ### Changed:
 - 9x10x9 model architecture